rework module path
